Add requests, pageable and test for cntrllrs, srvcs and vldtr

+++ src/main/java/ru/practicum/shareit/item/service/ItemServiceImpl.java
@@ -1,10 +1,13 @@
 package ru.practicum.shareit.item.service;
 
 import lombok.RequiredArgsConstructor;
-import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import ru.practicum.shareit.booking.mapper.BookingMapper;
+import ru.practicum.shareit.booking.model.Booking;
+import ru.practicum.shareit.booking.model.Status;
 import ru.practicum.shareit.booking.repository.BookingRepository;
 import ru.practicum.shareit.exception.*;
 import ru.practicum.shareit.item.dto.CommentDto;
@@ -15,6 +18,8 @@ import ru.practicum.shareit.item.model.Comment;
 import ru.practicum.shareit.item.model.Item;
 import ru.practicum.shareit.item.repository.CommentRepository;
 import ru.practicum.shareit.item.repository.ItemRepository;
+import ru.practicum.shareit.request.repository.ItemRequestRepository;
+import ru.practicum.shareit.user.model.User;
 import ru.practicum.shareit.user.repository.UserRepository;
 import ru.practicum.shareit.validator.CustomValidator;
 
@@ -22,8 +27,11 @@ import javax.transaction.Transactional;
 import java.time.LocalDateTime;
 import java.util.Collections;
 import java.util.List;
-import java.util.Optional;
-import java.util.stream.Collectors;
+import java.util.Map;
+
+import static java.util.Comparator.comparing;
+import static java.util.stream.Collectors.groupingBy;
+import static java.util.stream.Collectors.toList;
 
 @Slf4j
 @RequiredArgsConstructor
@@ -38,31 +46,43 @@ public class ItemServiceImpl implements ItemService {
 
     private final CommentRepository commentRepository;
 
+    private final ItemRequestRepository itemRequestRepository;
+
     private final CustomValidator customValidator;
 
-    @SneakyThrows
     @Transactional
     public ItemDto createItem(ItemDto itemDto, Long userId) {
         log.info("Попытка добавить новый предмет");
         customValidator.isItemValid(itemDto);
-        isUserExistAndAuthorizated(userId);
+        isUserAuthorizated(userId);
+        User user = userRepository.findById(userId).orElse(null);
+        if (user == null) {
+            log.info("Пользователь не зарегестрирован");
+            throw new ObjectNotFoundException("Пользователь не зарегестрирован");
+        }
         Item item = ItemMapper.toItem(itemDto);
-        item.setOwner(userRepository.getReferenceById(userId));
+        item.setOwner(user);
+        if (itemDto.getRequestId() != null) {
+            item.setRequest(itemRequestRepository.getReferenceById(itemDto.getRequestId()));
+        }
         log.info("Пользователь id: {} добавил новый предмет id: {}", userId, itemDto);
         return ItemMapper.toItemDto(itemRepository.save(item));
     }
 
-    @SneakyThrows
     @Transactional
     public ItemDto updateItem(ItemDto itemDto, Long itemId, Long userId) {
         log.info("Попытка обновить информацию о предмете id: {}", itemId);
-        isUserExistAndAuthorizated(userId);
-        isItemExist(itemId);
+        isUserAuthorizated(userId);
+        isUserExist(userId);
+        Item item = itemRepository.findById(itemId).orElse(null);
+        if (item == null) {
+            log.info("Предмет с id:{} не найден", itemId);
+            throw new ObjectNotFoundException("Такого предмета нет");
+        }
         if (!itemRepository.getOwnerId(itemId).equals(userId)) {
             log.info("Пользователь пытается обновить информацию о чужом предмете");
             throw new OwnerIdAndUserIdException("Обновлять информацию о предмете могут только владельцы");
         }
-        Item item = itemRepository.findById(itemId).get();
         if (itemDto.getName() != null) {
             item.setName(itemDto.getName());
         }
@@ -80,45 +100,60 @@ public class ItemServiceImpl implements ItemService {
     @Transactional
     public String deleteItem(Long itemId, Long userId) {
         log.info("Пользователь id: {} пытается удалить предмет id: {}", itemId);
-        isUserExistAndAuthorizated(userId);
+        isUserAuthorizated(userId);
+        isUserExist(userId);
         isItemExist(itemId);
+        if (!itemRepository.getOwnerId(itemId).equals(userId)) {
+            log.info("Пользователь пытается удалить чужой предмет");
+            throw new OwnerIdAndUserIdException("Удалять предметы могут только владельцы");
+        }
         itemRepository.deleteById(itemId);
         log.info("Предмет id: {} удален", itemId);
         return "Предмет удален";
     }
 
-    @SneakyThrows
     @Transactional
     public CommentDto addComment(CommentDto commentDto, Long itemId, Long userId) {
         log.info("Пользователь id: {} добавляет комментарий для предмета id: {}", userId, itemId);
-        isUserExistAndAuthorizated(userId);
-        isItemExist(itemId);
-        LocalDateTime createTime = LocalDateTime.now();
-        if (commentDto.getText().equals(null) || commentDto.getText().isEmpty()) {
-            log.info("Комментарий не содержит текст");
-            throw new EmptyCommentTextException("Текст комментария не может быть пустым");
+        isUserAuthorizated(userId);
+        User user = userRepository.findById(userId).orElse(null);
+        Item item = itemRepository.findById(itemId).orElse(null);
+        customValidator.isCommentValid(commentDto);
+        if (user == null) {
+            log.info("Пользователь не зарегестрирован");
+            throw new ObjectNotFoundException("Пользователь не зарегестрирован");
         }
-        if (bookingRepository.bookingsBeforeNowCount(itemId, userId, createTime) == 0) {
+        if (item == null) {
+            log.info("Предмет с id:{} не найден", itemId);
+            throw new ObjectNotFoundException("Такого предмета нет");
+        }
+        commentDto.setCreated(LocalDateTime.now());
+        if (bookingRepository.bookingsBeforeNowCount(itemId, userId, commentDto.getCreated()).equals(0L)) {
             log.info("Пользователь пока еще не брал в аренду предмет");
             throw new NoBookedYetException("Предмент не арендовался пользователем");
         }
         Comment comment = Comment.builder()
-                .author(userRepository.getReferenceById(userId))
-                .item(itemRepository.getReferenceById(itemId))
+                .author(user)
+                .item(item)
                 .text(commentDto.getText())
-                .created(createTime)
+                .created(commentDto.getCreated())
                 .build();
         log.info("Пользователь id: {} добавил комментарий для предмета id: {}", userId, itemId);
         return CommentMapper.toCommentDto(commentRepository.save(comment));
     }
 
-    @SneakyThrows
+    @Transactional
     public CommentDto updateComment(CommentDto commentDto, Long itemId, Long commentId, Long userId) {
         log.info("Пользователь id: {} пытается изменить комментарий id: {} для " +
                 "предмета id: {}", userId, commentId, itemId);
-        isUserExistAndAuthorizated(userId);
+        isUserAuthorizated(userId);
+        isUserExist(userId);
         isItemExist(itemId);
-        isCommentExist(commentId);
+        Comment comment = commentRepository.findById(commentId).orElse(null);
+        if (comment == null) {
+            log.info("Комментария id: {} не существует", commentId);
+            throw new ObjectNotFoundException("Такого комментария нет");
+        }
         if (!commentRepository.getCommentAuthorId(commentId).equals(userId)) {
             log.info("Пользователь id: {} не является автором комментария id: {}", userId, commentId);
             throw new OwnerIdAndUserIdException("Только авторы комментария могут его редактировать");
@@ -127,100 +162,135 @@ public class ItemServiceImpl implements ItemService {
             log.info("Получен пустое поле text, комментарий не обновлен");
             throw new EmptyCommentTextException("Невозможно сохранить пустой комментарий");
         }
-        Comment comment = commentRepository.getReferenceById(commentId);
         comment.setText(commentDto.getText());
         return CommentMapper.toCommentDto(commentRepository.save(comment));
     }
 
-    @SneakyThrows
+    @Transactional
     public String deleteComment(Long commentId, Long itemId, Long userId) {
         log.info("Пользователь id: {} пытается удалить комментарий id: {} для " +
                 "предмета id: {}", userId, commentId, itemId);
-        isUserExistAndAuthorizated(userId);
+        isUserAuthorizated(userId);
+        isUserExist(userId);
         isItemExist(itemId);
         isCommentExist(commentId);
+        if (!commentRepository.getCommentAuthorId(commentId).equals(userId)) {
+            log.info("Пользователь id: {} не является автором комментария id: {}", userId, commentId);
+            throw new OwnerIdAndUserIdException("Только авторы комментария могут его редактировать");
+        }
         log.info("Комментарий удален");
         commentRepository.deleteById(commentId);
         return "Комментарий удален";
     }
 
-    @SneakyThrows
     @Transactional
     public ItemDto findItemById(Long userId, Long itemId) {
         log.info("Попытка получить информацию о предмете id:{}", itemId);
-        isUserExistAndAuthorizated(userId);
-        isItemExist(itemId);
-        Item item = itemRepository.getReferenceById(itemId);
-        ItemDto itemDto = ItemMapper.toItemDto(itemRepository.getReferenceById(itemId));
-        if (commentRepository.isItemHaveComments(itemId) > 0) {
-            log.info("Получены все комментарии для предмета id: {}", itemId);
-            itemDto.setComments(getItemComments(itemId));
+        isUserAuthorizated(userId);
+        isUserExist(userId);
+        Item item = itemRepository.findById(itemId).orElse(null);
+        if (item == null) {
+            log.info("Предмет с id:{} не найден", itemId);
+            throw new ObjectNotFoundException("Такого предмета нет");
         }
-        itemDto.setComments(getItemComments(itemId));
+        List<Item> itemInList = List.of(item);
+        List<ItemDto> result = itemInList.stream().map(ItemMapper::toItemDto).collect(toList());
+        getCommentsForItem(itemInList, result);
         if (item.getOwner().getId().equals(userId)) {
             log.info("Получена информация о прошлом и следующем бронированиях предмета id: {}", itemId);
-            return getItemWithBookings(itemDto);
+            getBookingsForItem(itemInList, result);
         }
         log.info("Получена информация о предмете id: {}", itemId);
-        return itemDto;
+        return result.get(0);
     }
 
-    @SneakyThrows
     @Transactional
-    public List<ItemDto> findUserAllItems(Long userId) {
+    public List<ItemDto> findUserAllItems(Long userId, Integer from, Integer size) {
         log.info("Попытка получить список предметов пользователя {}", userId);
+        isUserAuthorizated(userId);
+        isUserExist(userId);
+        customValidator.isPageableParamsCorrect(from, size);
+        Pageable pageRequest = PageRequest.of(from, size);
+        List<Item> items = itemRepository.findAllByOwner_Id(userId, pageRequest);
+        List<ItemDto> result = items.stream().map(ItemMapper::toItemDto).collect(toList());
+        getCommentsForItem(items, result);
+        getBookingsForItem(items, result);
         log.info("Получен список предметов, размещенных пользователем {}", userId);
-        return itemRepository.findByOwner(userRepository.getReferenceById(userId)).stream()
-                .map(ItemMapper::toItemDto)
-                .map(this::getItemWithBookings)
-                .collect(Collectors.toList());
+        return result;
     }
 
-    private List<CommentDto> getItemComments(Long itemId) {
-        return commentRepository.getCommentsByItem_IdOrderByIdAsc(itemId).stream()
-                .map(CommentMapper::toCommentDto)
-                .collect(Collectors.toList());
+    private void getCommentsForItem(List<Item> items, List<ItemDto> result) {
+        if (!items.isEmpty()) {
+            Map<Item, List<Comment>> commentMap = commentRepository
+                    .findAllByItemInOrderByCreatedAsc(items)
+                    .stream()
+                    .collect(groupingBy(Comment::getItem, toList()));
+            for (int i = 0; i < result.size(); i++) {
+                if (!commentMap.containsKey(items.get(i))) {
+                    result.get(i).setComments(Collections.emptyList());
+                } else {
+                    result.get(i).setComments(commentMap.get(items.get(i)).stream()
+                            .map(CommentMapper::toCommentDto)
+                            .collect(toList()));
+                }
+            }
+        }
     }
 
-    private ItemDto getItemWithBookings(ItemDto itemDto) {
-        LocalDateTime nowTime = LocalDateTime.now();
-        itemDto.setLastBooking(BookingMapper
-                .toBookingForItemDto(Optional.ofNullable(bookingRepository.getLastBooking(nowTime, itemDto.getId()))
-                        .orElse(null)));
-        itemDto.setNextBooking(BookingMapper
-                .toBookingForItemDto(Optional.ofNullable(bookingRepository.getNextBooking(nowTime, itemDto.getId()))
-                        .orElse(null)));
-        return itemDto;
+
+    private void getBookingsForItem(List<Item> items, List<ItemDto> result) {
+        if (!items.isEmpty()) {
+            Map<Item, List<Booking>> bookingMap = bookingRepository
+                    .findAllByItemInAndStatusIsOrderByIdAsc(items,
+                            Status.APPROVED).stream()
+                    .collect(groupingBy(Booking::getItem, toList()));
+            for (int i = 0; i < result.size(); i++) {
+                if (bookingMap.containsKey(items.get(i))) {
+                    result.get(i).setLastBooking(bookingMap.get(items.get(i)).stream()
+                            .sorted(comparing(Booking::getStart).reversed())
+                            .filter(o -> o.getStart().isBefore(LocalDateTime.now()))
+                            .map(BookingMapper::toBookingForItemDto)
+                            .findFirst().orElse(null));
+                    result.get(i).setNextBooking(bookingMap.get(items.get(i)).stream()
+                            .sorted(comparing(Booking::getStart))
+                            .filter(o -> o.getStart().isAfter(LocalDateTime.now()))
+                            .map(BookingMapper::toBookingForItemDto)
+                            .findFirst().orElse(null));
+                }
+            }
+        }
     }
 
-    @SneakyThrows
     @Transactional
-    public List<ItemDto> searchItemByNameAndDescription(String text) {
+    public List<ItemDto> searchItemByNameAndDescription(String text, Integer from, Integer size) {
         log.info("Попытка получить список предметов, доступных для аренды, по запросу {}", text);
+        customValidator.isPageableParamsCorrect(from, size);
         if (text == null || text.isBlank()) {
             log.info("В поисковой строке пусто, получен пустой список предметов");
             return Collections.emptyList();
         }
+        Pageable pageRequest = PageRequest.of(from, size);
         log.info("Получен список предметов, доступных для аренды, по ключевому слову {}", text);
-        return itemRepository.searchByText(text).stream()
+        return itemRepository.searchByText(text, pageRequest).stream()
                 .map(ItemMapper::toItemDto)
                 .filter(o -> o.getAvailable() == true)
-                .collect(Collectors.toList());
+                .collect(toList());
     }
 
-    @SneakyThrows
-    private void isUserExistAndAuthorizated(Long userId) {
+    private void isUserAuthorizated(Long userId) {
         if (userId <= 0) {
             log.info("Пользователь не авторизован");
             throw new UserNotAuthorizedException("Пользователь не авторизован");
         }
+    }
+
+    private void isUserExist(Long userId) {
         if (!userRepository.existsById(userId)) {
             log.info("Пользователь не зарегестрирован");
             throw new ObjectNotFoundException("Пользователь не зарегестрирован");
         }
     }
 
-    @SneakyThrows
     private void isItemExist(Long itemId) {
         if (!itemRepository.existsById(itemId)) {
             log.info("Предмет с id:{} не найден", itemId);
@@ -228,7 +298,6 @@ public class ItemServiceImpl implements ItemService {
         }
     }
 
-    @SneakyThrows
     private void isCommentExist(Long commentId) {
         if (!commentRepository.existsById(commentId)) {
             log.info("Комментария id: {} не существует", commentId);
